flask_rest_with_sqlalchemy/requirements.txt
Flask_JWT
Flask_RESTful
Flask-SQLAlchemy